move poly fnction from app module to utils module
