+++ app.py
@@ -20,6 +20,7 @@ from torchvision.datasets import MNIST
 from torchvision.transforms import ToTensor
 
 from viz import mnist_like_viz, training_curves
+from utils import poly
 import dl
 
 
@@ -170,34 +171,6 @@ def regression(home_data):
     val_maes = pd.DataFrame(dict_val_maes).set_index("method")
     st.write(val_maes)
     st.text("(Test what happens when removing TotRmsAbvGrd)")
-
-
-def poly(x, order=3):
-    """Evaluates the different powers of an input vector.
-
-    The input vector is evaluated element-wise
-    to the power 1, 2, ..., `order`. The resulting vectors
-    are then concatenated and returned.
-
-    Parameters
-    ----------
-    x: array_like
-        The input vector, of shape `(n, 1)`.
-    order: int
-        The maximum order to which the powers of
-        `x`are computed.
-
-    Returns
-    -------
-    x_out: array_like
-        The concatenation of all
-        the powers of `x`, of shape `(n, order)`.
-
-    """
-    x_out = x
-    for i in range(2, order + 1):
-        x_out = np.concatenate((x_out, np.power(x, i)), axis=1)
-    return x_out
 
 
 def sinus():
